no funciona actualizar pendientes

+++ back/controllers/productController.js
@@ -144,9 +144,10 @@ const BookController = {
     },
     async pendientes (req, res){
         try{
-            const {id_libro, uid} = req.body;
+            const {id_libro, uid, estado} = req.body;
+            console.log('estado', estado + id_libro + uid)
             
-           /* if (actualizar) {
+            if (estado == 'leido') {
                
                 const deleteQuery = `DELETE FROM pendientes WHERE id_libro = "${id_libro}" AND uid = "${uid}"`;
                 await pool.query(deleteQuery);
@@ -157,18 +158,18 @@ const BookController = {
                 res.status(200).send('Libro movido de pendientes a leídos correctamente');
             } 
             else{
-                */
+                
                 const insertQuery = `INSERT INTO pendientes (id_libro, uid) VALUES ("${id_libro}", "${uid}")`;
                 const pendientes = await pool.query(insertQuery);
 
-                if (!pendientes) {
-                    console.log('Error al insertar el libro en la tabla por leer:', error);
+                if (pendientes) {
+                    console.log('Libro guardado como pendiente en la base de datos');
+                    res.status(201).send('Libro añadido como pendiente correctamente');
+                  } else {
+                    console.log('Error al insertar el libro en la tabla por leer:', result.error);
                     res.status(500).send('Error al guardar el libro como pendiente en la base de datos');
-                } else {
-                    console.log('libro guardado en la base de datos');
-                    res.status(201).send('añadido como pendiente correctamente');
-                }
-           // }
+                  }
+            }
         }
         catch(error){
             console.log(error)
